# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Smug
  description: 'Start (and never finish) side projects with this API SMUG Application.'
  version: 1.0.0
servers:
  -
    url: 'http://smug.test'
paths:
...
  /api/images/updateimage:
    post:
      summary: 'Update Image'
      operationId: updateImage
      description: 'This endpoint allows you to get update image .'
      parameters: []
      responses: {  }
      tags:
        - 'Image Manipulation'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                image:
                  type: string
                  description: .
                  example: et
              required:
                - image
...